Develop -> Master

+++ Machete.Util/sh/new-local-sqlserver-container.sh
@@ -1,4 +1,4 @@
-# Well aware this file is becoming kind of a nightmare. TODO cleanup.
+#!/bin/bash
 set +H
 
 if [ "${BASH_SOURCE[0]}" -ef "$0" ]
@@ -8,25 +8,19 @@ then
 fi
 
 sudo docker rm -f $MACHETE_SQL_DOCKER_CONTAINER
-
-if [ $? -ne 0 ]; then
-  sudo docker rm -f sqlserver
-fi
+if [ $? -ne 0 ]; then sudo docker rm -f sqlserver; fi
 
 sudo docker pull mcr.microsoft.com/mssql/server
-#sudo docker pull postgres
 
 if [[ $(cat /etc/hosts | grep sqlserver | wc -l) -eq 0 ]]; then
-#  sudo echo "127.0.0.1	sqlserver" >> /etc/hosts
    echo '127.0.0.1       sqlserver' | sudo tee -a /etc/hosts
 fi
 
-# this used to output the alread-running container name, it now does not :p so you'll have to delete manually if MACHETE_SQL_DOCKER_CONTAINER is not set.
-# to make your life easier, just copy-paste the value it gives you between quotes and `export MACHETE_SQL_DOCKER_CONTAINER={that}`
-export MACHETE_SQL_DOCKER_CONTAINER=$(sudo docker run -e 'ACCEPT_EULA=Y' -e 'SA_PASSWORD=passw0rD!' --network machete-bridge -p 1433:1433 --name sqlserver -d mcr.microsoft.com/mssql/server)
-#export MACHETE_SQL_DOCKER_CONTAINER=$(sudo docker run -e 'POSTGRES_PASSWORD=passw0rD!' --network machete-bridge -p 5432:5432 --name postgresql -d postgres)
+# output container name
+export MACHETE_SQL_DOCKER_CONTAINER=$(sudo docker run -e 'ACCEPT_EULA=Y' -e 'SA_PASSWORD=passw0rD!' --network bridge -p 1433:1433 --name sqlserver -d mcr.microsoft.com/mssql/server)
 echo $MACHETE_SQL_DOCKER_CONTAINER
 
+# Keep this for when we fix our dev permissions schema (corners were cut):
 # sudo docker exec -it sql1 /opt/mssql-tools/bin/sqlcmd -S localhost \
 #   -U SA -P 'passw0rD!' -Q $'CREATE USER dev WITH PASSWORD = \'passw0rD!\''
 # sudo docker exec -it sql1 /opt/mssql-tools/bin/sqlcmd -S localhost \
